% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{admm}
\alias{admm}
\title{admm to estimate some parameters}
\usage{
admm(beta0, beta1, beta2, m, n)
}
\arguments{
\item{beta0}{one parameter in a regression model}

\item{beta1}{one parameter in a regression model}

\item{beta2}{one parameter in a regression model}

\item{m}{the number of the sample}

\item{n}{the observation number for a sample}
}
\value{
a table and a number
}
\description{
estimate some parameters in a regression model by the admm method
}
\examples{
\dontrun{
admm(1.5,1,0,20,5)
} 
}
