% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df}
\alias{df}
\title{compute the df of cauchy distribution}
\usage{
df(x, theta, eta)
}
\arguments{
\item{x}{which the df take values in}

\item{theta}{the scale parameter}

\item{eta}{the location parameter}
}
\value{
a number
}
\description{
using MC integration
}
\examples{
\dontrun{
df(1,1,0)
} 
}
