% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est}
\alias{est}
\title{A esitimator for a parameter in a distribution}
\usage{
est(sigma, n)
}
\arguments{
\item{sigma}{the parameter to estimate}

\item{n}{the number of samples}
}
\value{
a random vector
}
\description{
estimate a parameter and compute its sd and bias
}
\examples{
\dontrun{
est(1,1e4)
} 
}
