% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC1}
\alias{gibbsC1}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC1(a, b, n)
}
\arguments{
\item{a}{related to the distribution}

\item{b}{related to the distribution}

\item{n}{the number of samples}
}
\value{
a random matrix
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
gibbsC1(1,2,10) 
} 
}
